requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pymupdf>=1.23.0
anthropic>=0.18.0
pydantic>=2.5.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
